+++ SourceryFramework/Sources/Parsing/SwiftSyntax/AST/GenericParameter+SwiftSyntax.swift
@@ -1,0 +1,9 @@
+import Foundation
+import SwiftSyntax
+import SourceryRuntime
+
+extension GenericParameter {
+    convenience init(_ node: GenericParameterSyntax) {
+        self.init(name: node.name.description.trimmed, inheritedTypeName: node.inheritedType.flatMap(TypeName.init(_:)))
+    }
+}